requirements.txt
plotly==5.17.0
kaleido==0.2.1
statsmodels==0.14.0
geopandas>=0.14.0
folium==0.14.0
fpdf2==2.7.4
python-docx==0.8.11
weasyprint==58.0
streamlit>=1.28,<1.30
pandas>=2.1,<2.3
numpy>=1.25,<1.27
pymupdf>=1.23,<1.25
pytesseract>=0.3.10,<0.4
Pillow>=10.0,<10.3
openpyxl>=3.1,<3.2
scikit-learn>=1.3,<1.4
xgboost>=1.7,<1.8
requests>=2.31,<2.32
python-multipart>=0.0.6,<0.1
imbalanced-learn>=0.11,<0.12
optuna>=3.3,<3.4
dateparser>=1.1,<1.2
fuzzywuzzy>=0.18,<0.19
python-levenshtein>=0.21,<0.22
joblib>=1.3,<1.4